Add block to show product specifications
A new block is added to the main-product section, which shows product
specifications.

There are two input settings for this block -
1. Section Heading - this allows user to enter a different heading for
the block than the default, which is "Product Specifications"
2. Metafield values - for this setting, the user has to input the names
of the product metafields that they want to show as table.
The values entered have to be comma separated,
Ex - Fabric,Color,Occasion,Length

Using the comma separated input values as keys, we get value of the
custom metafields and display them as a 2 column table